refactor(shortform): split compose and dispatch support
